+++ fj-doc-mod-fop/src/main/java/org/fugerit/java/doc/mod/fop/FopHelperConstants.java
@@ -1,0 +1,12 @@
+package org.fugerit.java.doc.mod.fop;
+
+public class FopHelperConstants {
+
+    private FopHelperConstants() {}
+
+    /**
+     * <a href="https://venusdocs.fugerit.org/docs/html/doc_meta_info.html#mod-fop-xslt-path">See 'mod-fop-xslt-path' documentation</a>
+     */
+    public static final String INFO_KEY_MOD_FOP_XSLT_PATH = "mod-fop-xslt-path";
+
+}