+++ fj-doc-mod-fop/src/main/java/org/fugerit/java/doc/mod/fop/PdfFopTypeHandler.java
@@ -7,6 +7,7 @@ import java.util.stream.Collectors;
 
 import javax.xml.transform.Result;
 import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerConfigurationException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.sax.SAXResult;
 import javax.xml.transform.stream.StreamSource;
@@ -26,6 +27,7 @@ import org.fugerit.java.core.util.mvn.MavenProps;
 import org.fugerit.java.core.xml.dom.DOMIO;
 import org.fugerit.java.core.xml.dom.DOMUtils;
 import org.fugerit.java.doc.base.config.*;
+import org.fugerit.java.doc.base.model.DocBase;
 import org.fugerit.java.doc.mod.fop.config.FopConfigClassLoaderWrapper;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
@@ -196,6 +198,18 @@ public class PdfFopTypeHandler extends FreeMarkerFopTypeHandler {
 		}
 	}
 
+	private Transformer newTransformer( TransformerFactory factory, DocBase docBase ) throws TransformerConfigurationException, IOException {
+		String xsltPath = docBase.getStableInfo().getProperty( FopHelperConstants.INFO_KEY_MOD_FOP_XSLT_PATH );
+		log.debug( "newTransformer {} -> {}", FopHelperConstants.INFO_KEY_MOD_FOP_XSLT_PATH, xsltPath );
+		if ( StringUtils.isNotEmpty( xsltPath ) ) {
+			try ( InputStream is = ClassHelper.loadFromDefaultClassLoader( xsltPath ) ) {
+				return factory.newTransformer( new StreamSource( is ) );
+			}
+		} else {
+			return factory.newTransformer();
+		}
+	}
+
 	@SuppressWarnings("unchecked")
 	@Override
 	public void handle(DocInput docInput, DocOutput docOutput) throws Exception {
@@ -205,7 +219,7 @@ public class PdfFopTypeHandler extends FreeMarkerFopTypeHandler {
 			FopConfigWrap fopWrap = this.fopWrapSupplier.get();
 			Fop fop = fopWrap.getFopFactory().newFop(MimeConstants.MIME_PDF, fopWrap.getFoUserAgent(), docOutput.getOs());
 			TransformerFactory factory = TransformerFactory.newInstance();
-			Transformer transformer = factory.newTransformer();
+			Transformer transformer = this.newTransformer( factory, docInput.getDoc() );
 			Result res = new SAXResult(fop.getDefaultHandler());
 			transformer.transform(xmlSource, res);
 			this.fopWrapConsumer.accept( fopWrap );