+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- TIFF validtor based on TwelveMonkeys ImageIO <https://github.com/fugerit-org/fj-doc/issues/551>
+
 ## [8.17.0] - 2025-10-14
 
 ### Added