+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Image validator based on TwelveMonkeys ImageIO (currently supporting TIFF) <https://github.com/fugerit-org/fj-doc/issues/551>
+- fj-doc-mod-fop, new mod-fop-xslt-path property <https://github.com/fugerit-org/fj-doc/issues/558>
 
 ### Changed