+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Image validator based on TwelveMonkeys ImageIO (currently supporting TIFF) <https://github.com/fugerit-org/fj-doc/issues/551>
-- fj-doc-mod-fop, new mod-fop-xslt-path property <https://github.com/fugerit-org/fj-doc/issues/558>
+- fj-doc-mod-fop, new <> property <https://github.com/fugerit-org/fj-doc/issues/558>
 
 ### Changed
 
 - quarkus-version set to 3.26.5 across all the modules
+- now DocInput has access to DocProcessContext <https://github.com/fugerit-org/fj-doc/issues/545>
 
 ### Fixed